+++ examples/simple_side_effect/src/aggregate.rs
@@ -80,7 +80,7 @@ impl LoggingAggregate {
 
         let policies: Vec<Box<dyn Policy<LoggingEvent, LoggingError> + Send + Sync>> = vec![Box::new(LoggingPolicy)];
 
-        PgStore::new::<Self>(pool, projectors, policies).await
+        PgStore::new::<Self>(pool, projectors, policies).setup().await
     }
 }