+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `Aggregate` should implement `name` function that act as `Identifier`.
 - `AggregateManager` should implement `EventStore` associated type.
 - `Projector` should implement `delete` function.
+- `PgStore::setup` function to create table and indexes if not exists.
 
 ### Changed
 
@@ -19,6 +20,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   order to handle persistence layer.
 - `Projector` second parameter changed from `Transaction` to `PgConnection`.
 - `PgStore` moved to `esrs::store::postgres` module.
+- `PgStore::new` function is now sync and its return value is no longer a `Result` but `Self`.
 - `Projector` moved to `esrs::store::postgres` module.
 - `Policy` moved to `esrs::store::postgres` module.
 - `Aggregate::apply_event` `payload` parameter changed from reference to value (`Self::Event`).
@@ -32,6 +34,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `ProjectorEraser` trait.
 - `EraserStore` trait.
 - `ProjectorStore` trait.
+- `Eraser` trait.
+- `AggregateState::new_with_state` removed due to potential inconsistency while loading state.
 
 ## [0.6.2]